mod to trunk to fix snapToStep in mbutils.
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@5208 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/lib_mbutil/MBUtils.cpp
@@ -1400,11 +1400,11 @@ float snapToStep(float gfloat, float step)
   if(step <= 0) 
     return(gfloat);
   float fval   = gfloat / step;    // Divide by step
-  int   itemp;
+  long int itemp;
   if(fval < 0.0)
-    itemp = (int)(fval-0.5);
+    itemp = (long int)(fval-0.5);
   else
-    itemp  = (int)(fval+0.5);      // round to nearest integer
+    itemp  = (long int)(fval+0.5);      // round to nearest integer
   fval = (float)itemp * step;      // multiply again by the step
   return(fval);
 }
@@ -1423,9 +1423,9 @@ float snapToStep(float gfloat, float step)
 float snapDownToStep(float gfloat, float step)
 { 
   if(step <= 0) return(gfloat);
-  float fval   = gfloat / step;    // Divide by step
-  int   itemp  = (int)(fval);      // round to nearest integer
-  fval = (float)itemp * step;      // multiply again by the step
+  float fval     = gfloat / step;    // Divide by step
+  long int itemp = (long int)(fval); // round to nearest integer
+  fval = (float)itemp * step;        // multiply again by the step
   return(fval);
 }
